Fix all cef binary targets
