+++ .run/build-artifacts-debug.run.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="build-artifacts-debug" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="run --bin build-artifacts" />
+    <option name="command" value="run --bin cef-artifacts" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <envs />
     <option name="emulateTerminal" value="true" />
