+++ .run/clean-all-debug.run.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="clean-all-debug" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="run --bin clean-all" />
+    <option name="command" value="run --bin cef-clean" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <envs />
     <option name="emulateTerminal" value="true" />
