+++ .run/clean-all-release.run.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="clean-all-release" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="run --bin clean-all --release" />
+    <option name="command" value="run --bin cef-clean --release" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <envs />
     <option name="emulateTerminal" value="true" />
