+++ crates/cef-ui-tools/src/bin/cef_build.rs
@@ -1,0 +1,8 @@
+use anyhow::Result;
+use cef_ui_tools::build;
+
+fn main() -> Result<()> {
+    cef_build()?;
+
+    Ok(())
+}
