+++ crates/cef-ui-tools/src/cef_build.rs
@@ -1,13 +1,10 @@
 use anyhow::Result;
-use std::{env::current_dir, fs::remove_dir_all, process::Command};
 use tracing::{level_filters::LevelFilter, subscriber::set_global_default, Level};
 use tracing_log::LogTracer;
 use tracing_subscriber::FmtSubscriber;
 
-/// This does the following:
-/// * Runs cargo clean.
-/// * Removes intermediate artifacts.
-pub fn clean_all() -> Result<()> {
+/// Build the project.
+pub fn cef_build() -> Result<()> {
     // This routes log macros through tracing.
     LogTracer::init()?;
 
@@ -17,17 +14,6 @@ pub fn clean_all() -> Result<()> {
         .finish();
 
     set_global_default(subscriber)?;
-
-    Command::new("cargo")
-        .args(&["clean"])
-        .output()?;
-
-    let artifacts_dir = current_dir()?.join("artifacts");
-
-    // Remove the artifacts directory.
-    if artifacts_dir.exists() {
-        remove_dir_all(&artifacts_dir)?;
-    }
 
     Ok(())
 }
