+++ crates/cef-ui-tools/src/lib.rs
@@ -1,7 +1,8 @@
-mod build_artifacts;
-mod clean_all;
+mod cef_artifacts;
+mod cef_build;
+mod cef_clean;
 mod util;
 
-pub use build_artifacts::*;
-pub use clean_all::*;
+pub use cef_artifacts::*;
+pub use cef_clean::*;
 pub use util::*;
